Refactor: 할일 카드 모달 컴포넌트 분리 작업 및 리팩토링
