+++ pages/index.tsx
@@ -1,8 +1,5 @@
 import LandingPage from './LandingPage';
 
-
-
-
 export default function Home() {
   return (
     <>
